Reimplement throw detection as it was faulty
When throwing a phone, there is a number of peaks and troughs.  The previous implementation would detect multiple throws in one, so the using Audio it would detect a number of throws and make the sound.  However, what’s worse, it would register a number of throws so your last throw was not necessarily your last throw i.e. it would detect 2-3 throws in one, and show the last throw, yet your first peak / trough may indicate a much bigger throw.  This was incredibly confusing.

So this implementation detects a typical throw curve with at least 3 peaks/troughs.